NopKernel (#6035)
* NopKernel

* fix

* move kernel_utils
