+++ oneflow/core/kernel/copy_hd_kernel.cpp
@@ -13,11 +13,24 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/copy_hd_kernel.h"
+#include "oneflow/core/kernel/kernel.h"
 
 namespace oneflow {
 
 #ifdef WITH_CUDA
+
+class CopyHdKernel final : public KernelIf<DeviceType::kGPU> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(CopyHdKernel);
+  CopyHdKernel() = default;
+  ~CopyHdKernel() = default;
+
+ private:
+  void ForwardDataContent(const KernelCtx&,
+                          const std::function<Blob*(const std::string&)>&) const override;
+  void ForwardHeader(const KernelCtx& ctx,
+                     const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override;
+};
 
 void CopyHdKernel::ForwardDataContent(
     const KernelCtx& ctx, const std::function<Blob*(const std::string&)>& BnInOp2Blob) const {
