+++ oneflow/core/kernel/foreign_input_kernel.cpp
@@ -13,12 +13,29 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/foreign_input_kernel.h"
+#include "oneflow/core/kernel/kernel.h"
 #include "oneflow/core/common/buffer_manager.h"
 #include "oneflow/core/register/ofblob.h"
 #include "oneflow/core/job/job_instance.h"
 
 namespace oneflow {
+
+class ForeignInputKernel final : public KernelIf<DeviceType::kCPU> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(ForeignInputKernel);
+  ForeignInputKernel() = default;
+  ~ForeignInputKernel() = default;
+
+  void Forward(const KernelCtx& ctx,
+               const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override {
+    ForwardDataContent(ctx, BnInOp2Blob);
+  }
+
+ private:
+  void ForwardDataContent(
+      const KernelCtx& ctx,
+      const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override;
+};
 
 void ForeignInputKernel::ForwardDataContent(
     const KernelCtx& ctx, const std::function<Blob*(const std::string&)>& BnInOp2Blob) const {
