+++ oneflow/core/kernel/foreign_watch_kernel.cpp
@@ -13,13 +13,27 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/foreign_watch_kernel.h"
+#include "oneflow/core/kernel/kernel.h"
 #include "oneflow/core/common/buffer_manager.h"
 #include "oneflow/core/common/protobuf.h"
 #include "oneflow/core/register/ofblob.h"
 #include "oneflow/core/job/foreign_watcher.h"
 
 namespace oneflow {
+
+template<DeviceType device_type>
+class ForeignWatchKernel final : public KernelIf<device_type> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(ForeignWatchKernel);
+  ForeignWatchKernel() = default;
+  ~ForeignWatchKernel() = default;
+
+ private:
+  bool IsStateless() const override { return false; }
+  void ForwardDataContent(
+      const KernelCtx& ctx,
+      const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override;
+};
 
 template<DeviceType device_type>
 void ForeignWatchKernel<device_type>::ForwardDataContent(
