+++ oneflow/core/kernel/output_kernel.cpp
@@ -13,12 +13,27 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/output_kernel.h"
+#include "oneflow/core/kernel/kernel.h"
 #include "oneflow/core/common/buffer_manager.h"
 #include "oneflow/core/job/job_instance.h"
 #include "oneflow/core/job/global_for.h"
 
 namespace oneflow {
+
+template<DeviceType device_type>
+class OutputKernel final : public KernelIf<device_type> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(OutputKernel);
+  OutputKernel() = default;
+  ~OutputKernel() = default;
+
+ private:
+  void ForwardDataContent(
+      const KernelCtx& ctx,
+      const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override;
+  void ForwardHeader(const KernelCtx& ctx,
+                     const std::function<Blob*(const std::string&)>& BnInOp2Blob) const override;
+};
 
 template<DeviceType device_type>
 void OutputKernel<device_type>::ForwardDataContent(
