+++ oneflow/core/kernel/shape_elem_cnt_kernel.cpp
@@ -13,9 +13,22 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/shape_elem_cnt_kernel.h"
+#include "oneflow/core/kernel/kernel.h"
 
 namespace oneflow {
+
+template<DeviceType device_type, typename T>
+class ShapeElemCntKernel final : public KernelIf<device_type> {
+ public:
+  OF_DISALLOW_COPY_AND_MOVE(ShapeElemCntKernel);
+  ShapeElemCntKernel() = default;
+  ~ShapeElemCntKernel() override = default;
+
+ private:
+  void ForwardDataContent(const KernelCtx&,
+                          const std::function<Blob*(const std::string&)>&) const override;
+  int32_t GetShapePartialElemCnt(const ShapeView& shape) const;
+};
 
 template<DeviceType device_type, typename T>
 void ShapeElemCntKernel<device_type, T>::ForwardDataContent(
