+++ oneflow/user/kernels/gather_kernel_util.cpp
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/gather_kernel_util.h"
+#include "oneflow/user/kernels/gather_kernel_util.h"
 
 namespace oneflow {
