+++ oneflow/user/kernels/gather_kernel_util.h
@@ -13,8 +13,8 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#ifndef ONEFLOW_CORE_KERNEL_GATHER_KERNEL_UTIL_H_
-#define ONEFLOW_CORE_KERNEL_GATHER_KERNEL_UTIL_H_
+#ifndef ONEFLOW_USER_KERNELS_GATHER_KERNEL_UTIL_H_
+#define ONEFLOW_USER_KERNELS_GATHER_KERNEL_UTIL_H_
 
 #include "oneflow/core/kernel/kernel_util.h"
 
@@ -37,4 +37,4 @@ struct GatherKernelUtilImpl final {
 
 }  // namespace oneflow
 
-#endif  // ONEFLOW_CORE_KERNEL_GATHER_KERNEL_UTIL_H_
+#endif  // ONEFLOW_USER_KERNELS_GATHER_KERNEL_UTIL_H_
