+++ oneflow/user/kernels/partial_fc_sample_kernel.cu
@@ -18,7 +18,7 @@ limitations under the License.
 #include "oneflow/core/framework/framework.h"
 #include "oneflow/core/kernel/new_kernel_util.h"
 #include "oneflow/core/common/balanced_splitter.h"
-#include "oneflow/core/kernel/gather_kernel_util.h"
+#include "oneflow/user/kernels/gather_kernel_util.h"
 #include "oneflow/core/common/not_equal_to_previous_adjacent_iterator.h"
 #include <cub/cub.cuh>
 #include <curand.h>
