+++ oneflow/user/kernels/square_sum_kernel_util.cu
@@ -13,7 +13,7 @@ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
 */
-#include "oneflow/core/kernel/square_sum_kernel_util.h"
+#include "oneflow/user/kernels/square_sum_kernel_util.h"
 #include "oneflow/core/cuda/atomic.cuh"
 #include <cub/cub.cuh>
